Some file and code maintenace. Moved files into folders and split code out into own files. Fixed FieldSettings model
