+++ src/apps/swcr/models/SwcrPackage.ts
